+++ src/open_refinery/__init__.py
@@ -220,7 +220,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "2.10.0"
+__version__ = "2.11.0"
 
 __all__ = [
     "Factory",